Incomplete TTS and Video Gen/requirements.txt
openai>=1.0.0
requests>=2.31.0
pathlib
typing
python-dotenv>=1.0.0
